+++ Sources/SwiftUIKit/Extensions/View+Enabled.swift
@@ -12,10 +12,10 @@ public extension View {
     
     /**
      Adds a condition that controls whether or not users can
-     interact with this view.
-     
-     This is the inverse of `disabled` which may be a little
-     easier to read for some conditions.
+     interact with this view. It's the inverse of `disabled`
+     and may be easier to read for some conditions.
      */
-    func enabled(_ condition: Bool) -> some View { disabled(!condition) }
+    func enabled(_ condition: Bool) -> some View {
+        disabled(!condition)
+    }
 }